feat: 탭 중앙 정렬 요청, 남은 인원 수 제거, 프로필 이미지 크기 키우기 (#386)
