+++ src/pages/class/components/TabWrapper/tabWrapper.css.ts
@@ -2,10 +2,10 @@ import { style } from '@vanilla-extract/css';
 import { vars } from '@/shared/styles/theme.css';
 
 export const tabPanelStyle = style({
-  paddingTop:"2.4rem",
-  paddingRight:"2rem",
-  paddingLeft:"2rem",
-  paddingBottom:"15rem",
+  paddingTop: '2.4rem',
+  paddingRight: '2rem',
+  paddingLeft: '2rem',
+  paddingBottom: '15rem',
   borderTop: '1px solid',
 
   borderColor: vars.colors.gray01,
